⚡️ 입력 메시지 상태를 Footer에서 관리하도록 변경
