+++ app/frontend/src/components/Sidebar/Contents/Chatting/ChattingFooter.tsx
@@ -1,21 +1,31 @@
+import { useState } from 'react';
+
 import { Button, Textarea } from '@/components';
 
 import * as styles from './index.css';
 
 type ChattingFooterProps = {
-  value: string;
-  onChange: (e: React.ChangeEvent<HTMLTextAreaElement>) => void;
-  onSubmit: (e: React.FormEvent<HTMLFormElement>) => void;
+  sendMessage: (message: string) => void;
 };
 
-export function ChattingFooter({
-  value,
-  onChange,
-  onSubmit,
-}: ChattingFooterProps) {
+export function ChattingFooter({ sendMessage }: ChattingFooterProps) {
+  const [message, setMessage] = useState('');
+
+  const onChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
+    setMessage(e.target.value);
+  };
+
+  const onSubmit = (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+    if (!message) return;
+
+    sendMessage(message);
+    setMessage('');
+  };
+
   return (
     <form className={styles.footer} onSubmit={onSubmit}>
-      <Textarea maxLength={300} fullWidth onChange={onChange} value={value} />
+      <Textarea maxLength={300} fullWidth onChange={onChange} value={message} />
       <Button
         type="submit"
         theme="primary"
